requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jwcrypto==1.5.0
numpy==1.25.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
